+++ sdk/src/main/java/com/mapbox/maps/renderer/gl/TextureRenderer.kt
@@ -60,9 +60,6 @@ internal class TextureRenderer(
     if (program == 0) {
       prepare()
     }
-    // Reset to guarantee widgets are drawn on top of map
-    GLES20.glDisable(GLES20.GL_STENCIL_TEST)
-
     GLES20.glUseProgram(program)
     GLES20.glBindBuffer(GLES20.GL_ARRAY_BUFFER, vbo[0])
     GLES20.glVertexAttribPointer(